requirements.unlocked.txt
dj-database-url
dj-search-url
django-cache-url
django-configurations
# Pending release
# @see https://github.com/datamade/django-councilmatic/pull/44
-e git://github.com/civictechto/django-councilmatic@toronto#egg=django-councilmatic
gunicorn
psycopg2
whitenoise
# Works with elasticsearch v1.x
elasticsearch>=1.0.0,<2.0.0
# Issue with elasticsearch 1.7
# @see https://github.com/django-haystack/django-haystack/pull/1268#issuecomment-192019280
django-haystack<=2.4.1
newrelic
